separate out diagnostics for execute/subscribe/resolve
now that the files are longer... :)
